add polling for f24
