+++ ts/features/messages/utils/strings.ts
@@ -1,3 +1,6 @@
+import { pipe } from "fp-ts/lib/function";
+import * as A from "fp-ts/lib/Array";
+import * as O from "fp-ts/lib/Option";
 import { PaymentNoticeStatus } from "@pagopa/io-app-design-system";
 import I18n from "../../../i18n";
 
@@ -19,3 +22,11 @@ export const getBadgeTextByPaymentNoticeStatus = (
       return "";
   }
 };
+
+export const getHeaderByKey = (headers: Record<string, string>, key: string) =>
+  pipe(
+    Object.entries(headers),
+    A.findFirst(([k, _]) => k.toLowerCase() === key),
+    O.map(([_, value]) => value),
+    O.toUndefined
+  );
